+++ src/platform/plugins/shared/controls/public/controls/data_controls/options_list_control/components/options_list_popover_suggestions.tsx
@@ -10,21 +10,15 @@
 import React, { useCallback, useEffect, useMemo, useRef, useState } from 'react';
 
 import type { UseEuiTheme } from '@elastic/eui';
-import { EuiFieldSearch, EuiHighlight, EuiSelectable, EuiSpacer, useEuiTheme } from '@elastic/eui';
+import { EuiHighlight, EuiSelectable, useEuiTheme } from '@elastic/eui';
 import type { EuiSelectableOption } from '@elastic/eui/src/components/selectable/selectable_option';
-import { UserAvatar, getUserDisplayName } from '@kbn/user-profile-components';
 import { useBatchedPublishingSubjects } from '@kbn/presentation-publishing';
 
 import { css } from '@emotion/react';
 import { useMemoCss } from '@kbn/css-utils/public/use_memo_css';
 import type { OptionsListSelection } from '@kbn/controls-schemas';
 import type { OptionsListSuggestions } from '../../../../../common/options_list/types';
-import { MAX_OPTIONS_LIST_REQUEST_SIZE, NO_ASSIGNEES_OPTION_KEY } from '../constants';
-import {
-  bulkGetUserProfilesWithAvatar,
-  getCachedUserProfileWithAvatar,
-  suggestUserProfilesWithAvatar,
-} from '../utils/user_profile_lookup';
+import { MAX_OPTIONS_LIST_REQUEST_SIZE } from '../constants';
 import { useOptionsListContext } from '../options_list_context_provider';
 import { OptionsListStrings } from '../options_list_strings';
 import { OptionsListPopoverEmptyMessage } from './options_list_popover_empty_message';
@@ -84,102 +78,9 @@ export const OptionsListPopoverSuggestions = ({
     [availableOptions, totalCardinality, showOnlySelected, allowExpensiveQueries]
   );
 
-  const isAssigneeField = fieldName === 'kibana.alert.workflow_assignee_ids';
-  const [assigneeSuggestedUids, setAssigneeSuggestedUids] = useState<string[]>([]);
-  const [assigneeProfilesVersion, setAssigneeProfilesVersion] = useState(0);
-  const [assigneeSuggestStatus, setAssigneeSuggestStatus] = useState<
-    'idle' | 'loading' | 'success' | 'error'
-  >('idle');
-  const [selectableOptions, setSelectableOptions] = useState<EuiSelectableOption[]>([]); // set in following useEffect
-
   const suggestions = useMemo<OptionsListSuggestions | OptionsListSelection[]>(() => {
-    // For assignees, reuse the action bar search input as a user search term.
-    if (isAssigneeField && searchString?.trim() && assigneeSuggestStatus === 'success') {
-      return assigneeSuggestedUids.map((uid) => ({ value: uid }));
-    }
     return (showOnlySelected ? selectedOptions : availableOptions) ?? [];
-  }, [
-    availableOptions,
-    selectedOptions,
-    showOnlySelected,
-    isAssigneeField,
-    searchString,
-    assigneeSuggestedUids,
-    assigneeSuggestStatus,
-  ]);
-
-  const visibleSelectableOptions = useMemo(() => {
-    if (!isAssigneeField) return selectableOptions;
-    const term = (searchString ?? '').trim().toLowerCase();
-    if (!term) return selectableOptions;
-
-    return selectableOptions.filter((option) => {
-      if (option.isGroupLabel) return true;
-      if (option.key === 'exists-option') return true;
-      if (option.key === NO_ASSIGNEES_OPTION_KEY) return true;
-
-      const label = typeof option.label === 'string' ? option.label : '';
-      const haystack = `${label} ${String(option.key ?? '')}`.toLowerCase();
-      return haystack.includes(term);
-    });
-  }, [isAssigneeField, searchString, selectableOptions]);
-
-  const suggestionUids = useMemo<string[]>(() => {
-    if (!isAssigneeField) return [];
-    return suggestions
-      .map((suggestion) => {
-        const value = typeof suggestion === 'object' ? suggestion.value : suggestion;
-        return String(value);
-      })
-      .filter((value) => Boolean(value) && value !== NO_ASSIGNEES_OPTION_KEY);
-  }, [isAssigneeField, suggestions]);
-
-  useEffect(() => {
-    if (!isAssigneeField) return;
-    const term = searchString?.trim() ?? '';
-    if (!term) {
-      setAssigneeSuggestedUids([]);
-      setAssigneeSuggestStatus('idle');
-      return;
-    }
-
-    let cancelled = false;
-    (async () => {
-      try {
-        setAssigneeSuggestStatus('loading');
-        const profiles = await suggestUserProfilesWithAvatar(term);
-        if (cancelled) return;
-        setAssigneeSuggestedUids(profiles.map((p) => p.uid));
-        setAssigneeProfilesVersion((v) => v + 1);
-        setAssigneeSuggestStatus('success');
-      } catch {
-        // best-effort: if user profiles suggest is unavailable, fall back to default suggestions
-        if (!cancelled) setAssigneeSuggestStatus('error');
-      }
-    })();
-
-    return () => {
-      cancelled = true;
-    };
-  }, [isAssigneeField, searchString]);
-
-  useEffect(() => {
-    if (!isAssigneeField) return;
-    let cancelled = false;
-
-    (async () => {
-      try {
-        await bulkGetUserProfilesWithAvatar(suggestionUids);
-        if (!cancelled) setAssigneeProfilesVersion((v) => v + 1);
-      } catch {
-        // best-effort: if security/user profiles are unavailable, fall back to default rendering
-      }
-    })();
-
-    return () => {
-      cancelled = true;
-    };
-  }, [isAssigneeField, suggestionUids]);
+  }, [availableOptions, selectedOptions, showOnlySelected]);
 
   const existsSelectableOption = useMemo<EuiSelectableOption | undefined>(() => {
     if (hide_exists || (!existsSelected && (showOnlySelected || suggestions?.length === 0))) return;
@@ -192,6 +93,7 @@ export const OptionsListPopoverSuggestions = ({
       'data-test-subj': 'optionsList-control-selection-exists',
     };
   }, [suggestions, existsSelected, showOnlySelected, hide_exists, styles]);
+  const [selectableOptions, setSelectableOptions] = useState<EuiSelectableOption[]>([]); // will be set in following useEffect
   useEffect(() => {
     /* This useEffect makes selectableOptions responsive to search, show only selected, and clear selections */
     const options: EuiSelectableOption[] = suggestions.map((suggestion) => {
@@ -200,54 +102,21 @@ export const OptionsListPopoverSuggestions = ({
         suggestion = { value: suggestion };
       }
 
-      const rawValue = String(suggestion.value);
-      const userProfile =
-        isAssigneeField && rawValue !== NO_ASSIGNEES_OPTION_KEY
-          ? getCachedUserProfileWithAvatar(rawValue)
-          : undefined;
-      const assigneeLabel =
-        userProfile && userProfile !== null ? getUserDisplayName(userProfile.user) : undefined;
-
-      const isNoAssignees = isAssigneeField && rawValue === NO_ASSIGNEES_OPTION_KEY;
-
       return {
-        key: rawValue,
-        label: String(
-          (isNoAssignees
-            ? OptionsListStrings.controlAndPopover.getNoAssignees()
-            : isAssigneeField
-            ? assigneeLabel
-            : undefined) ??
-            fieldFormatter(suggestion.value) ??
-            suggestion.value
-        ),
+        key: String(suggestion.value),
+        label: String(fieldFormatter(suggestion.value) ?? suggestion.value),
         checked: (selectedOptions ?? []).includes(suggestion.value) ? 'on' : undefined,
         'data-test-subj': `optionsList-control-selection-${suggestion.value}`,
         className:
           showOnlySelected && invalidSelections.has(suggestion.value)
             ? 'optionsList__selectionInvalid'
             : 'optionsList__validSuggestion',
-        prepend: isNoAssignees ? (
-          <UserAvatar size="s" />
-        ) : isAssigneeField && userProfile && userProfile !== null ? (
-          <UserAvatar user={userProfile.user} avatar={userProfile.data.avatar} size="s" />
-        ) : undefined,
         append:
           !showOnlySelected && suggestion?.docCount ? (
             <OptionsListPopoverSuggestionBadge documentCount={suggestion.docCount} />
           ) : undefined,
       } as EuiSelectableOption;
     });
-
-    // Special-case: allow filtering to “No assignees” even when `hideExists` is true.
-    if (isAssigneeField) {
-      options.unshift({
-        key: NO_ASSIGNEES_OPTION_KEY,
-        label: OptionsListStrings.controlAndPopover.getNoAssignees(),
-        checked: (selectedOptions ?? []).includes(NO_ASSIGNEES_OPTION_KEY) ? 'on' : undefined,
-        'data-test-subj': 'optionsList-control-selection-no-assignees',
-      });
-    }
 
     if (canLoadMoreSuggestions) {
       options.push({
@@ -275,8 +144,6 @@ export const OptionsListPopoverSuggestions = ({
     existsSelectableOption,
     canLoadMoreSuggestions,
     fieldFormatter,
-    isAssigneeField,
-    assigneeProfilesVersion,
     styles,
   ]);
 
@@ -324,28 +191,14 @@ export const OptionsListPopoverSuggestions = ({
 
   return (
     <>
-      {isAssigneeField && (
-        <>
-          <EuiFieldSearch
-            compressed
-            fullWidth
-            value={searchString ?? ''}
-            onChange={(event) => componentApi.setSearchString(event.target.value)}
-            data-test-subj="optionsList-assignee-search-input"
-            placeholder={OptionsListStrings.popover.getSearchPlaceholder('exact')}
-            aria-label={OptionsListStrings.popover.getSearchAriaLabel(fieldName)}
-          />
-          <EuiSpacer size="s" />
-        </>
-      )}
       <div data-test-subj="optionsList--scrollListener" ref={listRef}>
         <EuiSelectable
-          options={visibleSelectableOptions}
+          options={selectableOptions}
           renderOption={(option) => renderOption(option, searchString)}
           listProps={{ onFocusBadge: false }}
           aria-label={OptionsListStrings.popover.getSuggestionsAriaLabel(
             fieldName,
-            visibleSelectableOptions.length
+            selectableOptions.length
           )}
           emptyMessage={<OptionsListPopoverEmptyMessage showOnlySelected={showOnlySelected} />}
           onChange={(newSuggestions, event, changedOption) => {
